refactor: common-storage deps on common-meta-type instead of the reverse.
