+++ src/common/storage/Cargo.toml
@@ -14,6 +14,7 @@ common-arrow = { path = "../arrow" }
 common-auth = { path = "../auth" }
 common-base = { path = "../base" }
 common-exception = { path = "../exception" }
+common-meta-types = { path = "../../meta/types" }
 
 anyhow = { workspace = true }
 async-trait = "0.1"
