+++ src/common/storage/src/lib.rs
@@ -29,30 +29,12 @@
 //! - Intermediate data generated by query could be stored by temporary operator.
 
 #![allow(clippy::uninlined_format_args)]
-#![feature(no_sanitize)]
 #![feature(io_error_other)]
 
 mod config;
 pub use config::CacheConfig;
 pub use config::ShareTableConfig;
-pub use config::StorageAzblobConfig;
 pub use config::StorageConfig;
-pub use config::StorageFsConfig;
-pub use config::StorageFtpConfig;
-pub use config::StorageGcsConfig;
-pub use config::StorageHdfsConfig;
-pub use config::StorageHttpConfig;
-pub use config::StorageIpfsConfig;
-pub use config::StorageMokaConfig;
-pub use config::StorageObsConfig;
-pub use config::StorageOssConfig;
-pub use config::StorageParams;
-pub use config::StorageRedisConfig;
-pub use config::StorageS3Config;
-pub use config::STORAGE_FTP_DEFAULT_ENDPOINT;
-pub use config::STORAGE_GCS_DEFAULT_ENDPOINT;
-pub use config::STORAGE_IPFS_DEFAULT_ENDPOINT;
-pub use config::STORAGE_S3_DEFAULT_ENDPOINT;
 
 mod operator;
 pub use operator::init_operator;
@@ -64,11 +46,13 @@ pub use metrics::StorageMetrics;
 pub use metrics::StorageMetricsLayer;
 
 mod runtime_layer;
-mod utils;
 
 mod cache;
 pub use cache::FuseCachePolicy;
 
 mod column_node;
+
+mod stage;
 pub use column_node::ColumnNode;
 pub use column_node::ColumnNodes;
+pub use stage::init_stage_operator;
