+++ src/meta/app/Cargo.toml
@@ -14,7 +14,6 @@ test = false
 common-expression = { path = "../../query/expression" }
 common-meta-kvapi = { path = "../kvapi" }
 common-meta-types = { path = "../types" }
-common-storage = { path = "../../common/storage" }
 
 chrono = { workspace = true }
 enumflags2 = { version = "0.7.5", features = ["serde"] }
