+++ src/meta/proto-conv/src/config_from_to_protobuf_impl.rs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_meta_types::StorageFsConfig;
+use common_meta_types::StorageGcsConfig;
+use common_meta_types::StorageOssConfig;
+use common_meta_types::StorageS3Config;
 use common_protos::pb;
-use common_storage::StorageFsConfig;
-use common_storage::StorageGcsConfig;
-use common_storage::StorageOssConfig;
-use common_storage::StorageS3Config;
 
 use crate::reader_check_msg;
 use crate::FromToProto;
