+++ src/meta/proto-conv/src/table_from_to_protobuf_impl.rs
@@ -21,8 +21,8 @@ use chrono::DateTime;
 use chrono::Utc;
 use common_expression as ex;
 use common_meta_app::schema as mt;
+use common_meta_types::StorageParams;
 use common_protos::pb;
-use common_storage::StorageParams;
 
 use crate::reader_check_msg;
 use crate::FromToProto;
