+++ src/meta/proto-conv/src/user_from_to_protobuf_impl.rs
@@ -22,12 +22,12 @@ use std::convert::TryFrom;
 use chrono::DateTime;
 use chrono::Utc;
 use common_meta_types as mt;
+use common_meta_types::StorageFsConfig;
+use common_meta_types::StorageGcsConfig;
+use common_meta_types::StorageOssConfig;
+use common_meta_types::StorageParams;
+use common_meta_types::StorageS3Config;
 use common_protos::pb;
-use common_storage::StorageFsConfig;
-use common_storage::StorageGcsConfig;
-use common_storage::StorageOssConfig;
-use common_storage::StorageParams;
-use common_storage::StorageS3Config;
 use enumflags2::BitFlags;
 use num::FromPrimitive;
