+++ src/meta/proto-conv/tests/it/user_proto_conv.rs
@@ -20,6 +20,11 @@ use chrono::NaiveDateTime;
 use chrono::NaiveTime;
 use chrono::Utc;
 use common_meta_types as mt;
+use common_meta_types::StorageFsConfig;
+use common_meta_types::StorageGcsConfig;
+use common_meta_types::StorageOssConfig;
+use common_meta_types::StorageParams;
+use common_meta_types::StorageS3Config;
 use common_meta_types::UserIdentity;
 use common_meta_types::UserInfo;
 use common_meta_types::UserPrivilegeType;
@@ -27,11 +32,6 @@ use common_proto_conv::FromToProto;
 use common_proto_conv::Incompatible;
 use common_proto_conv::VER;
 use common_protos::pb;
-use common_storage::StorageFsConfig;
-use common_storage::StorageGcsConfig;
-use common_storage::StorageOssConfig;
-use common_storage::StorageParams;
-use common_storage::StorageS3Config;
 use enumflags2::make_bitflags;
 use pretty_assertions::assert_eq;
