+++ src/meta/proto-conv/tests/it/v012_table_meta.rs
@@ -19,7 +19,7 @@ use chrono::Utc;
 use common_expression as ce;
 use common_expression::types::NumberDataType;
 use common_meta_app::schema as mt;
-use common_storage::StorageParams;
+use common_meta_types::StorageParams;
 use maplit::btreemap;
 
 use crate::common;
