+++ src/meta/types/Cargo.toml
@@ -14,9 +14,7 @@ test = false
 common-exception = { path = "../../common/exception" }
 common-io = { path = "../../common/io" }
 common-meta-stoerr = { path = "../stoerr" }
-common-storage = { path = "../../common/storage" }
 
-opendal = { workspace = true }
 openraft = { workspace = true }
 
 anyerror = { workspace = true }
