+++ src/meta/types/src/lib.rs
@@ -14,6 +14,7 @@
 
 #![allow(clippy::uninlined_format_args)]
 #![feature(provide_any)]
+#![feature(no_sanitize)]
 
 //! This crate defines data types used in meta data storage service.
 
@@ -49,6 +50,7 @@ mod with;
 
 mod principal_identity;
 mod proto_display;
+mod storage_params;
 
 // reexport
 
@@ -163,6 +165,7 @@ pub use seq_num::SeqNum;
 pub use seq_value::IntoSeqV;
 pub use seq_value::KVMeta;
 pub use seq_value::SeqV;
+pub use storage_params::*;
 pub use tenant_quota::TenantQuota;
 pub use user_auth::AuthInfo;
 pub use user_auth::AuthType;
