+++ src/query/config/src/outer_v0.rs
@@ -26,19 +26,19 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_types::AuthInfo;
 use common_meta_types::AuthType;
+use common_meta_types::StorageAzblobConfig as InnerStorageAzblobConfig;
+use common_meta_types::StorageFsConfig as InnerStorageFsConfig;
+use common_meta_types::StorageGcsConfig as InnerStorageGcsConfig;
+use common_meta_types::StorageHdfsConfig as InnerStorageHdfsConfig;
+use common_meta_types::StorageMokaConfig as InnerStorageMokaConfig;
+use common_meta_types::StorageObsConfig as InnerStorageObsConfig;
+use common_meta_types::StorageOssConfig as InnerStorageOssConfig;
+use common_meta_types::StorageParams;
+use common_meta_types::StorageRedisConfig as InnerStorageRedisConfig;
+use common_meta_types::StorageS3Config as InnerStorageS3Config;
 use common_meta_types::TenantQuota;
 use common_storage::CacheConfig as InnerCacheConfig;
-use common_storage::StorageAzblobConfig as InnerStorageAzblobConfig;
 use common_storage::StorageConfig as InnerStorageConfig;
-use common_storage::StorageFsConfig as InnerStorageFsConfig;
-use common_storage::StorageGcsConfig as InnerStorageGcsConfig;
-use common_storage::StorageHdfsConfig as InnerStorageHdfsConfig;
-use common_storage::StorageMokaConfig as InnerStorageMokaConfig;
-use common_storage::StorageObsConfig as InnerStorageObsConfig;
-use common_storage::StorageOssConfig as InnerStorageOssConfig;
-use common_storage::StorageParams;
-use common_storage::StorageRedisConfig as InnerStorageRedisConfig;
-use common_storage::StorageS3Config as InnerStorageS3Config;
 use common_tracing::Config as InnerLogConfig;
 use common_tracing::FileConfig as InnerFileLogConfig;
 use common_tracing::StderrConfig as InnerStderrLogConfig;
