+++ src/query/management/tests/it/stage.rs
@@ -23,9 +23,9 @@ use common_meta_kvapi::kvapi::KVApi;
 use common_meta_types::SeqV;
 use common_meta_types::StageFile;
 use common_meta_types::StageParams;
+use common_meta_types::StorageParams;
+use common_meta_types::StorageS3Config;
 use common_meta_types::UserStageInfo;
-use common_storage::StorageParams;
-use common_storage::StorageS3Config;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_add_stage() -> Result<()> {
