+++ src/query/service/tests/it/storages/fuse/table_test_fixture.rs
@@ -38,10 +38,10 @@ use common_expression::TableField;
 use common_expression::TableSchemaRef;
 use common_expression::TableSchemaRefExt;
 use common_meta_app::schema::DatabaseMeta;
+use common_meta_types::StorageFsConfig;
+use common_meta_types::StorageParams;
 use common_sql::plans::CreateDatabasePlan;
 use common_sql::plans::CreateTablePlanV2;
-use common_storage::StorageFsConfig;
-use common_storage::StorageParams;
 use common_storages_fuse::FuseTable;
 use common_storages_fuse::FUSE_TBL_XOR_BLOOM_INDEX_PREFIX;
 use databend_query::interpreters::append2table;
