+++ src/query/service/tests/it/storages/system.rs
@@ -22,15 +22,15 @@ use common_expression::block_debug::pretty_format_blocks;
 use common_meta_types::AuthInfo;
 use common_meta_types::AuthType;
 use common_meta_types::RoleInfo;
+// use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
+use common_meta_types::StorageParams;
+use common_meta_types::StorageS3Config;
 use common_meta_types::UserGrantSet;
 use common_meta_types::UserInfo;
 use common_meta_types::UserOption;
 use common_meta_types::UserQuota;
 use common_metrics::init_default_metrics_recorder;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
-// use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
-use common_storage::StorageParams;
-use common_storage::StorageS3Config;
 use common_storages_system::BuildOptionsTable;
 use common_storages_system::CatalogsTable;
 use common_storages_system::ClustersTable;
