+++ src/query/sharing-endpoint/src/accessor.rs
@@ -14,8 +14,8 @@
 
 use common_base::base::GlobalInstance;
 use common_exception::Result;
+use common_meta_types::StorageParams;
 use common_storage::init_operator;
-use common_storage::StorageParams;
 use opendal::Operator;
 use time::Duration;
