+++ src/query/sql/src/planner/binder/location.rs
@@ -19,17 +19,17 @@ use std::io::Result;
 use anyhow::anyhow;
 use common_ast::ast::UriLocation;
 use common_config::GlobalConfig;
-use common_storage::StorageAzblobConfig;
-use common_storage::StorageFsConfig;
-use common_storage::StorageGcsConfig;
-use common_storage::StorageHttpConfig;
-use common_storage::StorageIpfsConfig;
-use common_storage::StorageOssConfig;
-use common_storage::StorageParams;
-use common_storage::StorageS3Config;
-use common_storage::STORAGE_GCS_DEFAULT_ENDPOINT;
-use common_storage::STORAGE_IPFS_DEFAULT_ENDPOINT;
-use common_storage::STORAGE_S3_DEFAULT_ENDPOINT;
+use common_meta_types::StorageAzblobConfig;
+use common_meta_types::StorageFsConfig;
+use common_meta_types::StorageGcsConfig;
+use common_meta_types::StorageHttpConfig;
+use common_meta_types::StorageIpfsConfig;
+use common_meta_types::StorageOssConfig;
+use common_meta_types::StorageParams;
+use common_meta_types::StorageS3Config;
+use common_meta_types::STORAGE_GCS_DEFAULT_ENDPOINT;
+use common_meta_types::STORAGE_IPFS_DEFAULT_ENDPOINT;
+use common_meta_types::STORAGE_S3_DEFAULT_ENDPOINT;
 use opendal::Scheme;
 use percent_encoding::percent_decode_str;
