+++ src/query/sql/tests/location.rs
@@ -23,18 +23,18 @@ use common_ast::ast::UriLocation;
 use common_base::base::GlobalInstance;
 use common_config::Config;
 use common_config::GlobalConfig;
-use common_sql::planner::binder::parse_uri_location;
-use common_storage::StorageFsConfig;
+use common_meta_types::StorageFsConfig;
 // use common_storage::StorageFtpConfig;
-use common_storage::StorageGcsConfig;
-use common_storage::StorageHttpConfig;
-use common_storage::StorageIpfsConfig;
-use common_storage::StorageOssConfig;
-use common_storage::StorageParams;
-use common_storage::StorageS3Config;
-use common_storage::STORAGE_GCS_DEFAULT_ENDPOINT;
-use common_storage::STORAGE_IPFS_DEFAULT_ENDPOINT;
-use common_storage::STORAGE_S3_DEFAULT_ENDPOINT;
+use common_meta_types::StorageGcsConfig;
+use common_meta_types::StorageHttpConfig;
+use common_meta_types::StorageIpfsConfig;
+use common_meta_types::StorageOssConfig;
+use common_meta_types::StorageParams;
+use common_meta_types::StorageS3Config;
+use common_meta_types::STORAGE_GCS_DEFAULT_ENDPOINT;
+use common_meta_types::STORAGE_IPFS_DEFAULT_ENDPOINT;
+use common_meta_types::STORAGE_S3_DEFAULT_ENDPOINT;
+use common_sql::planner::binder::parse_uri_location;
 
 #[test]
 fn test_parse_uri_location() -> Result<()> {
